Build camera tests as both 32 and 64 bit binaries.
Also fixes warnings and changes the build rule from
BUILD_EXECUTABLE to BUILD_NATIVE_TEST, so that the binary
doesn't get installed under /system/bin.

(cherry-picked from commit commit 550458b2e2ba280030d1774395f3062a3d9af01)

+++ camera/tests/camera_metadata_tests_fake_vendor.h
@@ -21,6 +21,8 @@
 #ifndef TESTING_CAMERA_METADATA_FAKEVENDOR_H
 #define TESTING_CAMERA_METADATA_FAKEVENDOR_H
 
+#include <stdint.h>
+
 #include <system/camera_metadata.h>
 #include <system/camera_vendor_tags.h>
 
@@ -70,11 +72,11 @@ const char *fakevendor_section_names[FAKEVENDOR_SECTION_COUNT] = {
     "com.fakevendor.scaler"
 };
 
-unsigned int fakevendor_section_bounds[FAKEVENDOR_SECTION_COUNT][2] = {
-    { FAKEVENDOR_SENSOR_START,          FAKEVENDOR_SENSOR_END },
-    { FAKEVENDOR_SENSOR_I_START,        FAKEVENDOR_SENSOR_I_END },
-    { FAKEVENDOR_COLORCORRECTION_START, FAKEVENDOR_COLORCORRECTION_END },
-    { FAKEVENDOR_SCALER_START,          FAKEVENDOR_SCALER_END}
+uint32_t fakevendor_section_bounds[FAKEVENDOR_SECTION_COUNT][2] = {
+    { (uint32_t) FAKEVENDOR_SENSOR_START,          (uint32_t) FAKEVENDOR_SENSOR_END },
+    { (uint32_t) FAKEVENDOR_SENSOR_I_START,        (uint32_t) FAKEVENDOR_SENSOR_I_END },
+    { (uint32_t) FAKEVENDOR_COLORCORRECTION_START, (uint32_t) FAKEVENDOR_COLORCORRECTION_END },
+    { (uint32_t) FAKEVENDOR_SCALER_START,          (uint32_t) FAKEVENDOR_SCALER_END}
 };
 
 vendor_tag_info_t fakevendor_sensor[FAKEVENDOR_SENSOR_END -
